requirements.txt
google-cloud-bigquery>=3.11.4
tqdm>=4.65.0 